% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox_function.R
\name{checkbox_function}
\alias{checkbox_function}
\title{Function for dealing with checkbox data}
\usage{
checkbox_function(data, aggregate = TRUE, metadata = NULL)
}
\arguments{
\item{metadata}{}
}
\value{
what does this return?
}
\description{
Function for dealing with checkbox data
}
