% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_function.R
\name{dropdown_function}
\alias{dropdown_function}
\title{Function for dealing with dropdown data}
\usage{
dropdown_function(data)
}
\arguments{
\item{data}{}
}
\value{
what does this return?
}
\description{
Function for dealing with dropdown data
}
