% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_function.R
\name{likert_function}
\alias{likert_function}
\title{Function for dealing with likert data}
\usage{
likert_function(data, aggregate = TRUE, metadata = NULL)
}
\arguments{
\item{data}{add info}

\item{aggregate}{add info}

\item{metadata}{add info}
}
\value{
what does this return?
}
\description{
Function for dealing with likert data
}
