% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_choice_function.R
\name{multiple_choice}
\alias{multiple_choice}
\title{Function for dealing with multiple choice data}
\usage{
multiple_choice(data, label = TRUE, aggregate = TRUE)
}
\arguments{
\item{aggregate}{}
}
\value{
what does this return?
}
\description{
Function for dealing with multiple choice data
}
