% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yes_no_function.R
\name{yes_no_function}
\alias{yes_no_function}
\title{Function for dealing with yes and no data}
\usage{
yes_no_function(data, pretty = TRUE, useNA = "ifany",
  aggregate = TRUE)
}
\arguments{
\item{data}{add info}

\item{pretty}{add info}

\item{aggregate}{add info}

\item{useNa}{add info}
}
\value{
what does this return?
}
\description{
Function for dealing with yes and no data
}
